client_max_body_size を unsigned long として保持

+++ srcs/config/location_conf.cpp
@@ -107,11 +107,11 @@ void LocationConf::AppendAllowedMethod(std::string method) {
   allowed_methods_.insert(method);
 }
 
-int LocationConf::GetClientMaxBodySize() const {
+unsigned long LocationConf::GetClientMaxBodySize() const {
   return client_max_body_size_;
 }
 
-void LocationConf::SetClientMaxBodySize(int client_max_body_size) {
+void LocationConf::SetClientMaxBodySize(unsigned long client_max_body_size) {
   client_max_body_size_ = client_max_body_size;
 }
